requirements.txt
pyroclient @ git+https://github.com/pyronear/pyro-api.git@main#egg=pkg&subdirectory=client
python-dotenv